Initial Commit of SQL Injection
The first commit in the SQL Injection Project.  This project is written
in Java using the Eclipse IDE (Juno Edition).  The project demonstrates
a simple Android database application that shows the manner in which
harmful code can be applied to a database application using SQL
Injection and one of the methods to defend against it (Using Parameters)